purchased items screen and check internet connection

+++ app/src/main/java/com/example/ecommerceapp/presentation/cart/CartScreen.kt
@@ -9,7 +9,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
@@ -161,7 +160,7 @@ fun CartScreen(navController: NavController, viewModel: CartViewModel = hiltView
                 scope.launch {
                     snackbarHostState.showSnackbar(
                         message = error.message ?: "Unknown Error",
-                        duration = SnackbarDuration.Long
+                        duration = SnackbarDuration.Short
                     )
                 }
             }
